Calculate header checksum correctly
The old header checksum code never worked. First it calculated checksum
incorrectly due to wrong algorithm, then we applied an incorrect fix in
2014 based on an incorrect GitHub comment.

Calculate the header checksum correctly as per community documentation:

* https://gbdev.io/pandocs/The_Cartridge_Header.html
* https://gbdev.gg8.se/wiki/articles/The_Cartridge_Header

The code is used multiple times so move it to a separate function to
remove duplication of code.

Tested by flashing good Redump ROMs to the cart and also verifying the
ROMs are correct with RGBDS's rgbfix.

Fixes: 8a1141b7e078 ("Correct header checksum, add debug when checksum OK")
Signed-off-by: Jamie Bainbridge <jamie.bainbridge@gmail.com>